Load code font after JavaScript bundle
Loading Operator Mono in the `<head>` meant we were blocking the first
render by quite a long time. The font had to be completely loaded before
the browser could construct the CSSOM, meaning until then nothing was
happening making our perceived loading time a lot slower than it needed
to be.

By moving the fonts `.css` file at the bottom of the `<body>` we have a
first render before that font is loaded, making perceived performance a
lot faster.

This does give us a FOUT (flash of unstyled text), BUT since it's only
the font we use for code it's unnoticable and/or very small.