[WIP] Add NavBar to homepage
This is a WIP idea, /cc @morajabi. The idea is that the navbar
background is transparent when the page is at the top, and it turns into
it's normal pink once the user scrolls.

This is WIP because even with debouncing it makes scrolling janky, and the
transition needs to be reworked to use opacity. (rather than background)
I don't have more time at the moment, so I'd be happy if somebody could
take this branch and finish it off!