Fix sidebar links by trimming output of elementToText
Sidebar was linking to https://www.styled-components.com/docs/advanced#server-side-rendering but the anchor in the doc itself rendered as https://www.styled-components.com/docs/advanced#server-side-rendering- (an extra dash at the end).

This was due to the heading having the tag ` | v2` appended, the string was being split on the pipe but the extra spaces weren't being trimmed.